refactor(cli): rename `ogx launch` command group to `ogx connect`
The `launch` terminology collided conceptually with starting the server.
Rename the command subgroup to `connect`, which better describes its
purpose of connecting OGX to external services like OpenCode.
