+++ docs/docs/references/ogx_cli_reference/index.md
@@ -31,7 +31,7 @@ You have two ways to install OGX:
 ## `ogx` subcommands
 
 1. `stack`: Allows you to build a stack using the `ogx` distribution and run a OGX server. You can read more about how to build a OGX distribution in the [Build your own Distribution](../../distributions/building_distro) documentation.
-2. `launch`: Launch third-party tools configured for OGX. Currently supports `opencode`. See [OpenCode Integration](../../building_applications/opencode_integration) for details.
+2. `connect`: Connect third-party tools to the running OGX server. Currently supports `opencode`. See [OpenCode Integration](../../building_applications/opencode_integration) for details.
 
 For downloading models, we recommend using the [Hugging Face CLI](https://huggingface.co/docs/huggingface_hub/guides/cli). See [Downloading models](#downloading-models) for more information.
 
@@ -42,7 +42,7 @@ llama --help
 ```
 
 ```text
-usage: ogx [-h] {stack,launch} ...
+usage: ogx [-h] {stack,connect} ...
 
 Welcome to the OGX CLI
 
@@ -50,10 +50,10 @@ options:
   -h, --help      show this help message and exit
 
 subcommands:
-  {stack,launch}
+  {stack,connect}
 
   stack                 Operations for the OGX / Distributions
-  launch                Launch third-party tools configured for OGX
+  connect               Connect third-party tools to the running OGX server
 ```
 
 ## Downloading models
