+++ docs/static/openai-coverage.json
@@ -1431,8 +1431,7 @@
                 {
                   "property": "POST.responses.200.content.application/json.properties.text",
                   "details": [
-                    "Type added: ['object']",
-                    "Default changed: None -> {'format': {'type': 'text'}}"
+                    "Type added: ['object']"
                   ]
                 },
                 {
