+++ src/ogx/cli/ogx.py
@@ -12,7 +12,7 @@ from ogx.log import setup_logging
 # Initialize logging early before any loggers get created
 setup_logging()
 
-from .launch import LaunchParser  # type: ignore[attr-defined]
+from .connect import ConnectParser  # type: ignore[attr-defined]
 from .letsgo import LetsGo
 from .run import Run
 from .stack import StackParser  # type: ignore[attr-defined]
@@ -39,7 +39,7 @@ class OGXCLIParser:
         LetsGo.create(subparsers)
         Run.create(subparsers)
         StackParser.create(subparsers)
-        LaunchParser.create(subparsers)
+        ConnectParser.create(subparsers)
 
         print_subcommand_description(self.parser, subparsers)
